refactor: 상품 댓글(comment) 디렉토리 제거 및 상품(product) 디렉토리 하위로 이동
