+++ kyumall-client/src/test/java/com/kyumall/kyumallclient/product/ProductCommentIntegrationTest.java
@@ -4,20 +4,20 @@ import static org.assertj.core.api.Assertions.*;
 
 import com.kyumall.kyumallclient.AuthTestUtil;
 import com.kyumall.kyumallclient.IntegrationTest;
-import com.kyumall.kyumallcommon.product.comment.dto.CreateCommentRequest;
-import com.kyumall.kyumallcommon.product.comment.dto.ProductCommentDto;
-import com.kyumall.kyumallcommon.product.comment.dto.UpdateCommentRequest;
+import com.kyumall.kyumallcommon.product.product.dto.CreateCommentRequest;
+import com.kyumall.kyumallcommon.product.product.dto.ProductCommentDto;
+import com.kyumall.kyumallcommon.product.product.dto.UpdateCommentRequest;
 import com.kyumall.kyumallcommon.factory.MemberFactory;
 import com.kyumall.kyumallcommon.factory.ProductFactory;
 import com.kyumall.kyumallcommon.fixture.member.MemberFixture;
 import com.kyumall.kyumallcommon.fixture.product.ProductFixture;
 import com.kyumall.kyumallcommon.member.entity.Member;
-import com.kyumall.kyumallcommon.product.product.Product;
-import com.kyumall.kyumallcommon.product.comment.ProductComment;
-import com.kyumall.kyumallcommon.product.comment.ProductCommentRating;
-import com.kyumall.kyumallcommon.product.comment.ProductCommentRatingRepository;
-import com.kyumall.kyumallcommon.product.comment.ProductCommentRepository;
-import com.kyumall.kyumallcommon.product.comment.RatingType;
+import com.kyumall.kyumallcommon.product.product.entity.Product;
+import com.kyumall.kyumallcommon.product.product.entity.ProductComment;
+import com.kyumall.kyumallcommon.product.product.entity.ProductCommentRating;
+import com.kyumall.kyumallcommon.product.product.repository.ProductCommentRatingRepository;
+import com.kyumall.kyumallcommon.product.product.repository.ProductCommentRepository;
+import com.kyumall.kyumallcommon.product.product.entity.RatingType;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
