+++ kyumall-common/src/main/java/com/kyumall/kyumallcommon/order/OrderService.java
@@ -10,10 +10,10 @@ import com.kyumall.kyumallcommon.order.entity.Orders;
 import com.kyumall.kyumallcommon.pay.PayService;
 import com.kyumall.kyumallcommon.order.entity.OrderItem;
 import com.kyumall.kyumallcommon.order.repository.OrdersRepository;
-import com.kyumall.kyumallcommon.product.product.Product;
+import com.kyumall.kyumallcommon.product.product.entity.Product;
 import com.kyumall.kyumallcommon.product.stock.ProductAndStockDto;
 import com.kyumall.kyumallcommon.product.stock.Stock;
-import com.kyumall.kyumallcommon.product.product.ProductRepository;
+import com.kyumall.kyumallcommon.product.product.repository.ProductRepository;
 import com.kyumall.kyumallcommon.product.stock.StockRepository;
 import jakarta.persistence.EntityManager;
 import java.time.LocalDateTime;
