--- kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/comment/RatingType.java
@@ -1,6 +1,0 @@
-package com.kyumall.kyumallcommon.product.comment;
-
-
-public enum RatingType {
-  LIKE, DISLIKE, NONE
-}
