--- kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/product/ProductStatus.java
@@ -1,5 +1,0 @@
-package com.kyumall.kyumallcommon.product.product;
-
-public enum ProductStatus {
-  INUSE, DELETED
-}
