+++ kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/product/entity/Product.java
@@ -1,4 +1,4 @@
-package com.kyumall.kyumallcommon.product.product;
+package com.kyumall.kyumallcommon.product.product.entity;
 
 import com.kyumall.kyumallcommon.BaseTimeEntity;
 import com.kyumall.kyumallcommon.member.entity.Member;
