+++ kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/product/entity/ProductComment.java
@@ -1,10 +1,8 @@
-package com.kyumall.kyumallcommon.product.comment;
+package com.kyumall.kyumallcommon.product.product.entity;
 
 import com.kyumall.kyumallcommon.BaseTimeEntity;
 import com.kyumall.kyumallcommon.member.entity.Member;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -18,24 +16,23 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter @AllArgsConstructor @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class ProductCommentRating extends BaseTimeEntity {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
+public class ProductComment extends BaseTimeEntity {
+  @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "product_comment_id")
-  private ProductComment productComment;
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "member_id")
   private Member member;
-  @Enumerated(EnumType.STRING)
-  private RatingType ratingType;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "product_id")
+  private Product product;
+  private String content;
+  @ManyToOne
+  @JoinColumn(name = "parent_id")
+  private ProductComment parentComment;
 
-  public void updateRating(RatingType ratingType) {
-    this.ratingType = ratingType;
+  public void updateComment(String newComment) {
+    this.content = newComment;
   }
 }
