+++ kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/product/entity/ProductCommentRating.java
@@ -1,9 +1,10 @@
-package com.kyumall.kyumallcommon.product.comment;
+package com.kyumall.kyumallcommon.product.product.entity;
 
 import com.kyumall.kyumallcommon.BaseTimeEntity;
 import com.kyumall.kyumallcommon.member.entity.Member;
-import com.kyumall.kyumallcommon.product.product.Product;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -17,23 +18,24 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Builder
-@Getter @AllArgsConstructor @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class ProductComment extends BaseTimeEntity {
-  @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+public class ProductCommentRating extends BaseTimeEntity {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "product_comment_id")
+  private ProductComment productComment;
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "member_id")
   private Member member;
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "product_id")
-  private Product product;
-  private String content;
-  @ManyToOne
-  @JoinColumn(name = "parent_id")
-  private ProductComment parentComment;
+  @Enumerated(EnumType.STRING)
+  private RatingType ratingType;
 
-  public void updateComment(String newComment) {
-    this.content = newComment;
+  public void updateRating(RatingType ratingType) {
+    this.ratingType = ratingType;
   }
 }
