+++ kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/product/entity/ProductStatus.java
@@ -1,0 +1,5 @@
+package com.kyumall.kyumallcommon.product.product.entity;
+
+public enum ProductStatus {
+  INUSE, DELETED
+}
