+++ kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/product/entity/RatingType.java
@@ -1,0 +1,6 @@
+package com.kyumall.kyumallcommon.product.product.entity;
+
+
+public enum RatingType {
+  LIKE, DISLIKE, NONE
+}
