+++ kyumall-common/src/main/java/com/kyumall/kyumallcommon/product/stock/StockRepository.java
@@ -1,15 +1,12 @@
 package com.kyumall.kyumallcommon.product.stock;
 
-import com.kyumall.kyumallcommon.product.product.Product;
-import com.kyumall.kyumallcommon.product.stock.Stock;
+import com.kyumall.kyumallcommon.product.product.entity.Product;
 import jakarta.persistence.LockModeType;
-import jakarta.persistence.QueryHint;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.jpa.repository.QueryHints;
 
 public interface StockRepository extends JpaRepository<Stock, Long> {
   Optional<Stock> findByProduct(Product product);
