+++ kyumall-common/src/testFixtures/java/com/kyumall/kyumallcommon/fixture/product/ProductCommentRatingFixture.java
@@ -1,9 +1,9 @@
 package com.kyumall.kyumallcommon.fixture.product;
 
 import com.kyumall.kyumallcommon.member.entity.Member;
-import com.kyumall.kyumallcommon.product.comment.ProductComment;
-import com.kyumall.kyumallcommon.product.comment.ProductCommentRating;
-import com.kyumall.kyumallcommon.product.comment.RatingType;
+import com.kyumall.kyumallcommon.product.product.entity.ProductComment;
+import com.kyumall.kyumallcommon.product.product.entity.ProductCommentRating;
+import com.kyumall.kyumallcommon.product.product.entity.RatingType;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
