+++ kyumall-common/src/testFixtures/java/com/kyumall/kyumallcommon/fixture/product/ProductFixture.java
@@ -2,7 +2,7 @@ package com.kyumall.kyumallcommon.fixture.product;
 
 import com.kyumall.kyumallcommon.member.entity.Member;
 import com.kyumall.kyumallcommon.product.category.Category;
-import com.kyumall.kyumallcommon.product.product.Product;
+import com.kyumall.kyumallcommon.product.product.entity.Product;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
